Allows to have separate DBs for GIS OSM info and rendering DB.
The assumption that the OSM postGIS DB was the same as the DB used for
the rendering (shared with maposmaticd) doesn't allow to run 2
instances of the maposmatic/maposmaticd engine (eg. a prod and a dev
instance). This patch allows this by differentiating both DB. This
makes it possible to share the same OSM db by both instances, but to
have separate rendering info.