Make sure than admin mode and bbox mode are really exclusive
When we had a submission in admin mode, we were correctly setting the
bbox coordinates to None. However, when we had a submission in bbox
mode, we didn't reset the administrative_city field to None. Which
means that we could have the following (common!) scenario:

 1. The user chooses the default mode, admin mode, and enters the name
    of a city

 2. MapOSMatic says "sorry, I don't know this city"

 3. The user switches to bbox mode, *without* blanking the
    administrative city field

 4. The user selects a bounding box and submits

In this case, we would save in the database both the administrative
city and the bounding box. Then, the maposmaticd daemon would choose
the administrative city... and fail because the administrative city
doesn't exist.

This commit probably fixes the problem that caused most of the
rendering failures in bounding box mode.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@free-electrons.com>