Query nominatim for more details + canonicalize the results
With the "addressdetail" querty param, Nominatim allows to retrieve
the coutry, world region, etc. as well as the language code. The
latter is needed for localization. This patch uses it and correctly
pareses the new format. The patch also transforms the results so that
dict, list, etc. do not only deal with strings: we try to cast the
strings to integers, float, etc. as far as possible.