FIXED bug with IE: force javastcript to encode the nominatim query arguments
When the city query contains special features (accents, etc.), IE was
getting the nominatim query URL without escaping them. This patch adds
the ability for the nominatim pages to be queried through GET. We then
ask jQuery to pass the query arguments through GET, which has the nice
consequence that everything is escaped correctly.