Revamp the jobs and maps pages
This is a major revamp of the user experience on the jobs and maps
listing pages, and job detailled view page.

Jobs and rendered maps are no longer displayed in a table that could
easily overflow small screens. A list of jobs is used instead, with
nice, visual information on its status.

An automatic 5s refresh was also added to the individual job status page
when the job is not completed.