Cornerstones for a new MapOSMatic daemon
This is the first step of the rewrite of the MapOSMatic rendering
daemon. The long-term objective is to get a more flexible rendering
daemon that would support the rendering of several jobs in parallel, and
with an overall more Python-y and maintainable code.

This first shot brings a completely new, simpler MapOSMatic daemon with
the same level of functionality as the previous daemon in terms of
serial rendering of the job queue. Three major changes happen here:

  1. the bash-based wrapper script has been removed, in favor of a more
     clever Python wrapper. Cleaner, and more importantly more portable.
     The wrapper still needs a bit of configuration, and after the
     config.py-template has been tweaked into a valid config.py file,
     the daemon can be started by:

       .../scripts/wrapper.py scripts/daemon.py

  2. the externalization of the rendering routine into a 'render'
     module. This module does not access the database - only the daemon
     does. The sole purpose of the render module is to encapsulate the
     rendering process and its errors+exceptions handling.

     It can also be used as a standalone, job-ID-based renderer:

       .../scripts/wrapper.py scripts/render.py <jobid>

  3. the cleanup mechanism now runs in a separate thread, in the
     background.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>