Revamp the job renderer module with timeout support
The render module, originally a simple extract of the render_job
function from the old daemon, has been rewriten to expose a more
sensible API and to provide self-contained timeout support.

Two classes are now available, exposing the same public API for seamless
use of any of the two:

  * JobRenderer is a simple, blocking job renderer. It can be threaded
    if the caller decides to call start() instead of run().
  * TimingOutJobRenderer, as its name suggests, is a timing out job
    renderer that makes use of the threading capability of JobRenderer
    to handle a timeout on the rendering process and kill the rendering
    thread, whatever it is doing, if the given timeout is reached.

The render module also now exposes a few public RESULT_ constants that
can be used to identify the result of a job rendering. This is used in
the daemon to infer an appropriate resultmsg.

As a standalone process, the job renderer now takes an optional second
argument: the timeout, in seconds.

  .../scripts/wrapper.py scripts/render.py <jobid> [timeout]

As of now, the daemon is fully usable in production with the same level
of functionality as before.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>