Fix the language selection dropdown
When switching to a locale-specific language on the MapOSMatic website
with the language selection dropdown, the new language was not made the
current selection of the dropdown menu, and fell back to the default
'French'. It was thus impossible to go back to French without switch to
an intermediary language.

The behavior was the result of a difference between our (traditional)
language codes and the ones used by Django. The language list has been
fixed, and a comment added to proof for future additions.