Add a rendered maps feed
This change introduces a RSS feeds for the successfully rendered maps
that still have their files available (status=2) from the last 24 hours
(or more to get at least 10 items when possible), using the Django
syndication framework.

It uses a custom map-feed.html template instead of the map.html in order
to include styling information directly in the HTML an accomodate for
the variable naming from the syndication framework.

This new MapOSMatic maps feed is now available from a traditional RSS
icon link on the maps page.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>