Make sure the byadmin and bybbox mode are really exclusive
This fixes task #10150. When selecting a city by administrative
boundary, and the rendering is refused because the area is too large,
and then switching to bounding box mode trying to render an acceptable
area, the big administrative boundary is rendered instead. This is
because of two things:

  * the administrative city and administrative osmid fields are not
    cleared when switching to bounding box / when cleaning the form data
    for bounding box mode
  * the daemon renders the given osmID even if a bounding box is
    provided.

This changes fixes the first issue, which effectively resolves task
10150. The second issue will be fixed by the new (upcoming) daemon code.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>