Fall back to the last 10 maps in the RSS feed
When no map has been rendered in the last 24 hours, the RSS feed can be
empty, which is not good. Fall back to the last 10 entries, regardless
of any time limit, when this happens.

If no map at all was ever rendered, well we don't have much choice but
to send back an empty feed.

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>