Job cancel feature
This change introduces the job cancel feature as requested in bug #28488
(https://savannah.nongnu.org/bugs/?28488).

To do so, a new field called 'nonce' is added to the MapRenderingJob
model object (which results in a new column in the database, see below
how to recreate/update a current database). A random string, called the
nonce, is generated when the job is created and stored alongside the
other information on this job. The user is then redirected to the
/jobs/ID/NONCE page, instead of simply /jobs/ID.

As long as the user provides the correct nonce at the end of the URL, he
will have the ability, if the job is still in the queue, to cancel the
job request. The button to cancel the request is shown if and only if
the user provides the matching nonce string in the URL, which is only
displayed when the job is created.

A new job state is also created (4, Cancelled) to match this new
cancelled state.

As far as the database change goes, the easiest if you are only doing
development is to drop your database and recreate it with syncdb. If you
need to keep your data, you can simply add the column to the
maposmatic_maprenderingjob table using the following statement:

  ALTER TABLE maposmatic_maprenderingjob ADD COLUMN
    nonce varchar(16) NOT NULL;

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>
Acked-by: Thomas Petazzoni <thomas.petazzoni@enix.org>