Add a get_bbox_from_osm_id() helper
To implement the /papersize/ service, we need to get the bbox
associated to a given OSM id, so this commit adds a
get_bbox_from_osm_id() helper.

In fact, the code for this helper already existed in
check_osm_id(). So we've moved part of check_osm_id() into
get_bbox_from_osm_id(), which is now used by check_osm_id().

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>
Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>