Add layout, stylesheet and papersize to MapRenderingForm
The form used to ask the user for the various parameters of the map
needs to be extended with the new layout, stylesheet and papersize
parameters.

As OCitySMap doesn't yet implement the list of layouts, stylesheets
and paper sizes, so we use stub lists.

The ChoiceField choice lists are filled dynamically in the __init__
constructor instead of directly using the "choices" argument of the
ChoiceField constructor so that the result of a function call can be
used as the list of possible choices.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>
Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>