Use the reversegeo service to find the country code
The callback that the slippy map calls when an area is selected is
modified so that it gets as argument 'bounds', which represents the
bounding box of the selected area.

Using this, we get the center of the bounding box, and using an Ajax
request, ask the reversegeo service to find the corresponding
entries. When the reply is received, we extract the country code from
the received reply.

This will be used in a later commit to pre-filter the list of
languages according to the selected geographic area.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>
Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>