Improve the language selector
Instead of removing languages that do not match the current country
code, we instead order the languages in the following way :

 1. All languages that match the current country code

 2. A separator

 3. All languages that do not match the current country code

This allows to have a sane choice selected by default, to highlight
the choices that are probably the most pertinent ones, and still give
the user the opportunity to make a completely different choice. This
is especially important for bounding box renderings, since the guessed
country code may not be correct (in particular for bounding boxes
crossing country boundaries).

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>
Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>