Fix paper_width_mm/paper_height_mm handling
Introduce two hidden fields in MapRenderingJobForm to store
paper_width_mm and paper_height_mm. Those two fields are updated by
the JS code when the selected paper size changes. They are then sent
to the server as part of the wizard form and then stored in the
database.

It allows to fix the handling of paper sizes for "Best fit"
renderings.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>
Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>