Fixed form cleaning behavior upon error
When an error occurs while processing the form's data, cleaning up the
data can result in attempting to delete non existent keys. This patch
should make the code tolerant to the half-created form data we have to
deal with in that case.

The problem happened 5 or 6 times within the last 3 weeks.