The concept of "Index" file has changed
With ocitysmap 1.x, an "index" file used to be a file whose name ended
in "_index.*". With ocitysmap2, the index is generally included in the
map image (depending on the chosen page layout). But we still generate
the index separately as a CSV file. This patch simply tells that the
index file is the CSV.