Language sorting change
On the website, the language selector dropdown list was ordered in a
more or less random way, depending on how languages were listed in the
MAP_LANGUAGES array of the settings.py file. This had the unfortunate
effect of proposing a language which is not necessarily the most
common language for the country in which the city to be rendered
is. For example, for all cities located in France, the language
proposed was Catalan (because ca_FR is before fr_FR in the
alphabetically-sorted list of locales).

Therefore, this commit does two things to solve this problem:

 * The MAP_LANGUAGES array in settings.py is now statically sorted by
   country rather by language, and then the different languages of a
   given country must be manually sorted (by the developer who adds
   them) from the most widely used language in the country to the
   least widely used language.

 * The Javascript code is modified so that the order of the languages
   in MAP_LANGUAGES is preserved. For this, a new jQuery function
   called reverse() has been added.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>