maposmatic/newmap: implement message when no admin boundaries are available
Nominatim returns some search results that are not appropriate for
MaOSMatic because they do not have an administrative boundary. We show
those results so that users aren't confused, but we also display a
message explaining why those results can't be used. This commit
implements this message.

Instead of putting the message in the Javascript, which would require
interfacing the Javascript code with the Django translation/locale
system, the message is stored in an hidden <span> element in the HTML
code (and is therefore translated like all the rest of the HTML text
by the Django translation infrastructure). This message is copied by
the Javascript code at the right place when needed. We were already
doing the same thing for the #noresultsinfo message.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>