maposmatic/nominatim: fix Nominatim handling of next results
Contrary to what we initially thought, there is no way to know if
Nominatim has further results or not. Even if the first results
returned by Nominatim are limited to 5 results (i.e less than the
maximum number of results Nominatim can return on a single page), it
does not mean that Nominatim has no further results.

So, we just assume that Nominatim may have further results.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>