Show MapOSMatic GIS database lag with the official OSM database
This commit implements a new item on the right of the MapOSMatic
website, which gives users the current time delta between our GIS
database and the official OSM database.

In order to implement this, we created a small gisdb.py module which
factorizes the work of connecting to the GIS database, now used by
both the Nominatim code and our new code that fetches the last update
of the GIS database through the maposmatic_admin table.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>