Add small infrastructure to translate the Paypal stuff
We want the Paypal donation button to be translated, and the user to
be redirected to a page in his/her language when making a donation.

Unfortunately, Paypal does not seem to have translations for all the
languages we support, so we default to English when the language is
not available from Paypal.

This new mechanism requires translators to add a new entry in the
PAYPAL_LANGUAGES dictionary if Paypal provides a button and a site in
their language.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>