Adjust "No index" message for multi-page layout
We display a message saying that an index has not been generated if no
CSV was found, and in this case we encourage users to contribute to
OpenStreetMap. However, with the multi-page layout, we never generate
a CSV, so this message was always displayed. This commit handles this
case properly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>