models: extend can_recreate() to take into account new fields
All the jobs that are part of our database and have been submitted
with the previous MapOSMatic version have their layout, stylesheet,
paper_width_mm and paper_height_mm set to default values (empty for
string fields, -1 for integer fields), because they didn't exist in
the previous MapOSMatic version.

The problem is that when an user clicks on a "Recreate this map" link,
we re-create a new job with the same properties... and it is now
invalid to have a job with layout, stylesheet empty or unset
paper_width_mm and paper_height_mm.

Therefore, we adjust can_recreate() to ensure that it is not possible
to recreate jobs from jobs that have been rendered with the old
version of MapOSMatic. We could have set layout and stylesheet to
default values when they are empty, but it was much more complicated
to set a sane value for paper_width_mm and paper_height_mm.

We hope this is one of the reason why we have new jobs with the
stylesheet and layout fields empty, and the paper_width_mm and
paper_height_mm fields set to -1.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>