Rework configuration templates
Use the more traditional '.dist' extension for configuration templates
that are part of the MapOSMatic distribution. Update the installation
instructions accordingly.

Includes some style fixes in the INSTALL and template files, as well as
a fix to the wrapper script to correctly allow for an undefined log
location (defaults to stderr).

Signed-off-by: Maxime Petazzoni <maxime.petazzoni@bulix.org>