map.html: do not allow recreation of maps when it is not possible
The map.html contains a button that allows to recreate a similar map
to the one currently being shown. However, this button should only be
made visible if the map can be re-created, i.e if it has all the
necessary informations to do so.

This patch might fix the "Request stylesheet not found" errors we were
seeing regularly.

Signed-off-by: Thomas Petazzoni <thomas.petazzoni@enix.org>